Fix system package dependency
For the vendor package code to find a system package for libyaml, it
needs yaml-0.1.pc, which is provided by the libyaml-dev rosdep key (and
the libyaml-dev package on Debian/Ubuntu).

This change also adds a build dependency for all downstream packages
which take a build dependency on this one as yaml-0.1.pc will be needed
there too.

Signed-off-by: Scott K Logan <logans@cottsay.net>